+++ libup2p/rlpx_frame.h
@@ -44,7 +44,7 @@ int rlpx_frame_write(rlpx_coder* x,
                      uint8_t* data,
                      size_t datalen,
                      uint8_t* out,
-                     size_t* l);
+                     uint32_t* l);
 int rlpx_frame_parse(rlpx_coder* x, const uint8_t* frame, size_t l, urlp**);
 
 #ifdef __cplusplus
